requirements.txt
gspread==6.1.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
requests==2.31.0
PyYAML==6.0.1
